refactor: extract tools and resources into sub-servers
Extract computational tools (calculate, statistics, compound_interest,
convert_units) and persistence tools (save_calculation, load_variable)
into dedicated FastMCP sub-servers using the mount() pattern.

Extract resources (constants, functions, history, workspace) and
prompts (math_tutor, formula_explainer) into a separate resources
sub-server.

Changes:
- Create src/math_mcp/tools/calculate.py with 4 computational tools
- Create src/math_mcp/tools/persistence.py with 2 persistence tools
- Create src/math_mcp/resources.py with resources and prompts
- Update src/math_mcp/tools/__init__.py to export new sub-servers
- Update src/math_mcp/server.py to mount new sub-servers
- Update test imports to reference new module locations

Result: Reduced server.py from 773 to 193 lines while maintaining
all functionality. All 141 tests pass with clean linting.
